Fixes in lost calculation

+++ Throttle/ThrComms.h
@@ -56,8 +56,8 @@ class ThrComms {
       command.value = 0;
     }
 
-    int getLostRate() {
-      int lostRate = 0;
+    uint16_t getLostRate() {
+      uint16_t lostRate = 0;
       if (total) {
         lostRate = 100 * lost / total;
         if (lostRate > 100)
